Update and rename fahrenheit_to_celsius.py to temperature_conversion.py
Added a new function to convert celius to fahrenheit
